Move SensitiveErrorTextRedactor to Clio.Common (#1375)
Since #1333/#1372/#1374 the redactor is the product-wide untrusted-text rule:
Clio.Common.ClassifyingDataProvider, SysSettingsManager and
ExceptionReadableMessageExtension call it on the plain CLI path, not only the MCP
transport. Living in namespace Clio.Command.McpServer therefore made the shared
foundation layer import a transport-specific module while the redactor imported
Clio.Common back - a two-way namespace dependency in the wrong direction. The
class doc already named the move as deliberately deferred.

The type now sits in clio/Common/ next to TextUtilities, whose IsDisplayHostile /
NeutralizeDisplayHostileCharacters it builds on. Consumers under
Clio.Command.McpServer gained "using Clio.Common;" (KnowledgeSourceManagementService
takes a using alias instead, because importing Clio.Common wholesale there makes
IFileSystem ambiguous with System.IO.Abstractions.IFileSystem); clio/Common/ and
clio/Package/ dropped their "using Clio.Command.McpServer;". The test fixture moved
to clio.tests/Common with Module=Common and its entry in test-shards.json was
renamed with it.

No behaviour change: redaction rules, fence tokens, limits and timeouts are
byte-identical. Doc updates only remove the now-false "deferred move" sentences in
the class doc, UntrustedText and the knowledge record, and repoint the applies-to
paths of the two knowledge records that named the old file.

Validated: dotnet test clio.tests/clio.tests.csproj --filter "Category=Unit&(Module=Common|Module=McpServer|Module=Command)"
Failing-test set identical to the origin/master baseline (25 pre-existing
macOS-only failures, 10786 passed, 10917 total on both).


AI agents: none
