+++ clio/Command/McpServer/Knowledge/KnowledgeReferenceExampleService.cs
@@ -1,3 +1,4 @@
+using Clio.Common;
 using System;
 using System.Globalization;
 using System.Collections.Generic;
