+++ clio/Command/McpServer/McpToolErrorFilter.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Clio.Common;
+using System;
 using System.ComponentModel.DataAnnotations;
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
