+++ clio/Command/McpServer/Tools/ComponentInfoResolution.cs
@@ -1,3 +1,4 @@
+using Clio.Common;
 using System;
 using System.Threading.Tasks;
