+++ clio/Command/McpServer/Tools/DataForgeEnrichmentBuilder.cs
@@ -1,3 +1,4 @@
+using Clio.Common;
 using System;
 using System.Collections.Generic;
 using System.Linq;
