+++ clio/Command/McpServer/Tools/ExportComponentRegistryTool.cs
@@ -1,3 +1,4 @@
+using Clio.Common;
 using System;
 using System.Collections.Generic;
 using System.ComponentModel;
